Add CSS styles for footer content and image block
Included styles.css file to define the visual appearance of the footer content and image block. Applied styles to the #footerContent element to create a flexible layout using flexbox, allowing items to wrap and adjusting their alignment. Also, styled the h2 and span elements within the #footerContent for consistent line height. Added padding to the div elements within the #footerContent for spacing. Additionally, defined styles for the .image-block class to align images to the right and set their height. These styles ensure proper layout and presentation of the footer content and image block.