Add media query for mobile responsiveness
Included a media query in styles.css to enhance mobile responsiveness for screen widths up to 600 pixels. Within the media query, adjusted the order of part3 and part5 elements to swap their positions, ensuring a more intuitive layout for smaller screens. Hid part1 to optimize space on mobile devices. Modified the layout of part2-container, part2, and part5 to occupy the full width of the screen. Additionally, made adjustments to the width and float properties of part3 and part4 within the #footerContent to improve the layout for mobile viewing. These changes ensure a more user-friendly experience on mobile devices.