Enhance mobile responsiveness with media query and additional styles
Incorporated a media query in styles.css targeting screen widths up to 600 pixels to optimize the layout for mobile devices. Within the media query, swapped the positions of part3 and part5, hid part1 to conserve space, and adjusted the layout of part2-container, part2, and part5 to occupy the full width of the screen. Additionally, made adjustments to the width and float properties of part3 and part4 within the #footerContent to improve the layout for mobile viewing.
 included additional styles for the image block (#flexibleFooter .image-block) to center the images and adjust spacing between them. Added styles for the footer text (#flexibleFooter .footer-text) below the image block to ensure proper alignment and spacing. These enhancements contribute to a more user-friendly experience on mobile devices.